Everything is now part of caseStatus
