+++ lib/caseStatus.spec.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import * as caseStatusAPI from './caseStatus';
-import { mockedCaseStatusFactory } from 'factories/caseStatus';
+import { mockedCaseStatusFactory, mockedFormValue } from 'factories/caseStatus';
 
 const ENDPOINT_API = process.env.ENDPOINT_API;
 const AWS_KEY = process.env.AWS_KEY;
@@ -26,6 +26,28 @@ describe('case status APIs', () => {
         'x-api-key': AWS_KEY,
       });
       expect(data).toEqual(caseStatus);
+    });
+  });
+});
+
+describe('case form values APIs', () => {
+  describe('GetFormValues', () => {
+    it("calls the service API's form values endpoint", async () => {
+      mockedAxios.get.mockResolvedValue({
+        data: mockedFormValue,
+      });
+
+      const data = await caseStatusAPI.GetFormValues('CIN');
+
+      expect(mockedAxios.get).toHaveBeenCalled();
+      expect(mockedAxios.get.mock.calls[0][0]).toEqual(
+        `${ENDPOINT_API}/case-status/form-options/CIN`
+      );
+      expect(mockedAxios.get.mock.calls[0][1]?.headers).toEqual({
+        'x-api-key': AWS_KEY,
+      });
+
+      expect(data).toEqual(mockedFormValue);
     });
   });
 });
